Added LRU
Not tested due to faulty internet.
Methodology:
Iterates through the file with a while loop. Every time the loop iterates the int count is increased. Whenever a frame is used the value of this integer is assigned to a different array(recent) at the exact position of the modified frame. Instead of the int evict being randomly assigned a for loop will go through the entire recent array, find the smallest integer and set the value of evict to that integer.